+++ CHANGELOG.md
@@ -75,3 +75,4 @@
 * Work around KT-65315 thanks to updated conventions
 * BigNum as API dependency and iOS export (seems nonsensical,
   but is somehow required when using this inside a compose multiplatform app)
+* Rename `BERTags.NULL` to `BERTags.ASN1_NULL` to fix broken ObjC export